fix: Use full card menu everywhere (fix #3993)
